Changes needed to build on Windows
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@756144 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/decaf/util/concurrent/atomic/AtomicReferenceTest.cpp
@@ -134,7 +134,7 @@ void AtomicReferenceTest::testGetAndSet() {
 void AtomicReferenceTest::testToString() {
     int value;
     AtomicReference<int> ai;
-    CPPUNIT_ASSERT( ai.toString() == Integer::toString( 0 ) );
+    CPPUNIT_ASSERT( ai.toString() == Long::toString( 0 ) );
     ai.set( &value );
-    CPPUNIT_ASSERT( ai.toString() == Integer::toString( (long long)&value ) );
+    CPPUNIT_ASSERT( ai.toString() == Long::toString( (long long)&value ) );
 }
